requirements.txt
streamlit
python-dotenv
groq
pinecone-client
